+++ modules/cachix.nix
@@ -1,9 +1,7 @@
 {
-  nixosModule = { ... }: {
-    nix.settings.substituters = [ "https://enzime.cachix.org" ];
-    nix.settings.trusted-public-keys = [
-      "enzime.cachix.org-1:RvUdpEy6SEXlqvKYOVHpn5lNsJRsAZs6vVK1MFqJ9k4="
-    ];
+  nixosModule = { inputs, ... }: {
+    nix.settings.substituters = inputs.self.outputs.nixConfig.extra-substituters;
+    nix.settings.trusted-public-keys = inputs.self.outputs.nixConfig.extra-trusted-public-keys;
   };
 
   hmModule = { pkgs, ... }: {